added post security

+++ routes/posts.js
@@ -3,29 +3,41 @@ import PostModel from '../models/posts.js';
 
 const router = express.Router();
 
+const removeCreatorID = (posts) => {
+    return posts.map((post) => {
+        return {title: post.title, description: post.description, date: post.date,
+            _id: post.id, image: post.image, username: post.username, tags: post.tags, likes: post.likes, comments: post.comments}
+    })
+}
+
 router.get('/', async (req, res) => {
     try {
         const posts = await PostModel.find();
-        res.status(200).json(posts);
+        res.status(200).json(removeCreatorID(posts));
     } catch (error) {
         res.status(404).json({ message: error.message });
     }
 });
 
-router.get('/:id', async (req, res) => {
+router.get('/:creatorID', async (req, res) => {
     try {
-        const post = await PostModel.find({creatorID: req.params.id});
-        res.status(200).json(post);
+        const posts = await PostModel.find({creatorID: req.params.creatorID});
+        res.status(200).json(removeCreatorID(posts));
     } catch (error) {
         res.status(404).json({ message: error.message });
     }
 });
 
 router.delete('/:id', async (req, res) => {
+    const { creatorID } = req.body;
+
     try {
         const post = await PostModel.findById(req.params.id);
+        if(post.creatorID !== creatorID) {
+            return res.status(400).send("you are not authorized to delete this post")
+        }
         post.delete()
-        res.status(200).json(post);
+        res.status(200).send("deleted successfully");
     } catch (error) {
         res.status(404).json({ message: error.message });
     }
@@ -46,8 +58,8 @@ router.patch('/like/:id', async (req, res) => {
     const { id } = req.params;
     try {
         const post = await PostModel.findById(id);
-        const updatedPost = await PostModel.findByIdAndUpdate(id, { likes: post.likes + 1 }, { new: true });
-        res.json(updatedPost);
+        await PostModel.findByIdAndUpdate(id, { likes: post.likes + 1 }, { new: true });
+        res.json("liked successfully");
     } catch (error) {
         res.status(409).json({ message: error.message });
     }
@@ -55,8 +67,11 @@ router.patch('/like/:id', async (req, res) => {
 
 router.patch('/edit/:id', async (req, res) => {
     const { id } = req.params;
-    const { title, description, tags, image } = req.body;
+    const { title, description, tags, image, creatorID } = req.body;
     try {
+        if(post.creatorID !== creatorID) {
+            return res.status(400).send("you are not authorized to edit this post")
+        }
         const updatedPost = await PostModel.findByIdAndUpdate(id, { title, description, tags, image }, { new: true });
         res.json(updatedPost);
     } catch (error) {
